# docs/swagger.yaml
definitions:
  bizmandu.Sector:
    properties:
      label:
        type: string
      value:
        type: number
    type: object
  bizmandu.Topstock:
    properties:
      noOfStocks:
        type: number
      ticker:
        type: string
      totalHolding:
        type: number
    type: object
  bizmandu.Topstockholdings:
    properties:
      portWeight:
        type: number
      qty:
        type: integer
      ticker:
        type: string
    type: object
  controllers.FloorsheetResult:
    properties:
      buyerAveragePriceMap:
        additionalProperties:
          type: number
        type: object
      buyerQuantityMap:
        additionalProperties:
          type: number
        type: object
      buyerTurnOverMap:
        additionalProperties:
          type: number
        type: object
      sellerAveragePriceMap:
        additionalProperties:
          type: number
        type: object
      sellerQuantityMap:
        additionalProperties:
          type: number
        type: object
      sellerTurnOverMap:
        additionalProperties:
          type: number
        type: object
      ticker:
        type: string
    type: object
  controllers.MutualFund:
    properties:
      mutualFundKeyMetrics:
        items:
          $ref: '#/definitions/controllers.MutualFundKeyMetrics'
        type: array
      netStockBoughtMap:
        additionalProperties:
          type: number
        type: object
      sectorMap:
        additionalProperties:
          type: number
        type: object
      topHoldingMap:
        additionalProperties:
          type: number
        type: object
      topStockBoughtMap:
        additionalProperties:
          type: number
        type: object
      topStockSoldMap:
        additionalProperties:
          type: number
        type: object
    type: object
  controllers.MutualFundKeyMetrics:
    properties:
      lastTradedPrice:
        type: number
      marketCapatilization:
        type: number
      monthlyNav:
        type: number
      priceVsNav:
        type: number
      sector:
        items:
          $ref: '#/definitions/bizmandu.Sector'
        type: array
      ticker:
        type: string
      topStockBought:
        items:
          $ref: '#/definitions/bizmandu.Topstock'
        type: array
      topStockHoldings:
        items:
          $ref: '#/definitions/bizmandu.Topstockholdings'
        type: array
      topStockSold:
        items:
          $ref: '#/definitions/bizmandu.Topstock'
        type: array
      totalCompanies:
        type: integer
      totalSector:
        type: integer
      weeklyNav:
        type: number
    type: object
  controllers.TransactionData:
    properties:
      brokerPercentageShare:
        type: number
      percentageShare:
        type: number
      quantity:
        type: integer
      ticker:
        type: string
      totalQuantity:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/v1/floorsheet:
    get:
      consumes:
      - application/json
      description: Get the floorsheet of single stock (Pumper, Dumper, HODLers) in
        html format using go-echarts
      parameters:
      - description: start
        in: query
        name: start
        required: true
        type: string
      - description: end
        in: query
        name: end
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: string
      - description: ticker
        in: query
        name: ticker
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.FloorsheetResult'
      summary: Floorsheet of single the stocks between start and end date
      tags:
      - floorsheet
  /api/v1/floorsheet/analysis:
    get:
      consumes:
      - application/json
      description: Get Pumped/Dumped Stock Top Broker Buy/Sell in html format using
        go-echarts
      parameters:
      - description: start
        in: query
        name: start
        required: true
        type: string
      - description: end
        in: query
        name: end
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Analysis the Floorsheet of all the stocks between start and end date
      tags:
      - floorsheet
  /api/v1/floorsheet/bulk:
    get:
      consumes:
      - application/json
      description: Get Pumped/Dumped Stock Top Broker Buy/Sell in html format using
        go-echarts(aggregated)
      parameters:
      - description: start
        in: query
        name: start
        required: true
        type: string
      - description: end
        in: query
        name: end
        required: true
        type: string
      - description: id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/controllers.TransactionData'
              type: array
            type: object
      summary: Analysis the Floorsheet of all the stocks between start and end date
      tags:
      - floorsheet
  /api/v1/fundamental:
    get:
      consumes:
      - application/json
      description: This endpoint will create Sectorwise file with key indicators representating
        certain sectors so that we can compare the performance of the company in same
        sector
      parameters:
      - description: sector
        in: query
        name: sector
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create the fundamental report of the company sectorwise
      tags:
      - fundamental
  /api/v1/mutualfund:
    get:
      consumes:
      - application/json
      description: This endpoint will create the Mutual Funds reports about holding,
        bought and sold stocks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.MutualFund'
      summary: Create the Mutual Funds reports about holding, bought and sold stocks
      tags:
      - mutualfund
swagger: "2.0"